requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
charset-normalizer==2.0.8
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.3
Django==3.2.9
django-cors-headers==3.10.0
django-environ==0.8.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
filelock==3.3.2
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
kavenegar==1.1.2
MarkupSafe==2.0.1
numpy==1.21.4
openapi-codec==1.3.2
opencv-python==4.5.4.60
Pillow==8.4.0
platformdirs==2.4.0
psycopg2==2.9.2
PyJWT==2.3.0
pytz==2021.3
requests==2.26.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.10.0